backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
websockets==12.0
pydantic==2.5.3
python-dotenv==1.0.0
asyncio==3.4.3
typing-extensions==4.9.0
openai==1.12.0